Conflicts:
	LICENSE
	enforcer-ng/src/protobuf-orm/test/Makefile.am

+++ signer/src/scheduler/fifoq.c
@@ -1,6 +1,4 @@
 /*
- * $Id$
- *
  * Copyright (c) 2011 NLNet Labs. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -33,7 +31,7 @@
 
 #include "config.h"
 #include "scheduler/fifoq.h"
-#include "shared/log.h"
+#include "log.h"
 
 #include <ldns/ldns.h>
 
@@ -103,7 +101,10 @@ fifoq_pop(fifoq_type* q, worker_type** worker)
     }
     q->count -= 1;
     if (q->count <= (size_t) FIFOQ_MAX_COUNT * 0.1) {
-        /* notify waiting workers that they can start queuing again */
+        /**
+         * Notify waiting workers that they can start queuing again
+         * If no workers are waiting, this call has no effect.
+         */
         lock_basic_broadcast(&q->q_nonfull);
     }
     return pop;
@@ -121,7 +122,11 @@ fifoq_push(fifoq_type* q, void* item, worker_type* worker, int* tries)
         return ODS_STATUS_ASSERT_ERR;
     }
     if (q->count >= FIFOQ_MAX_COUNT) {
-        /* #262 if drudgers remain on hold, do additional broadcast */
+        /**
+         * #262:
+         * If drudgers remain on hold, do additional broadcast.
+         * If no drudgers are waiting, this call has no effect.
+         */
         if (*tries > FIFOQ_TRIES_COUNT) {
             lock_basic_broadcast(&q->q_threshold);
             ods_log_debug("[%s] queue full, notify drudgers again", fifoq_str);
@@ -134,8 +139,9 @@ fifoq_push(fifoq_type* q, void* item, worker_type* worker, int* tries)
     q->owner[q->count] = worker;
     q->count += 1;
     if (q->count == 1) {
-        ods_log_deeebug("[%s] threshold %u reached, notify drudgers",
-            fifoq_str, q->count);
+        ods_log_deeebug("[%s] threshold %lu reached, notify drudgers",
+            fifoq_str, (unsigned long) q->count);
+        /* If no drudgers are waiting, this call has no effect. */
         lock_basic_broadcast(&q->q_threshold);
     }
     return ODS_STATUS_OK;
